+++ packages/api/cms-api/src/dam/dam.config.ts
@@ -1,3 +1,5 @@
+export const damDefaultBasePath = "dam";
+
 export interface DamConfig {
     secret: string;
     allowedImageSizes: number[];
