+++ packages/api/cms-api/src/dam/dominant-color-calculator.interface.ts
@@ -1,0 +1,3 @@
+export interface DominantColorCalculatorInterface {
+    calculateDominantColor(contentHash: string): Promise<string | undefined>;
+}
